Added update (no incremental)

+++ genjax/distributions/beta.py
@@ -15,14 +15,13 @@
 import jax
 import jax.numpy as jnp
 from jax._src import abstract_arrays
-from jax._src import dtypes
 
 
 class Beta:
     def abstract_eval(self, key, a, b, shape=None):
         return (
             key,
-            abstract_arrays.ShapedArray(shape=shape, dtype=dtypes.float_),
+            abstract_arrays.ShapedArray(shape=shape, dtype=jnp.float32),
         )
 
     def sample(self, key, a, b, **kwargs):
