+++ genjax/distributions/cauchy.py
@@ -15,14 +15,13 @@
 import jax
 import jax.numpy as jnp
 from jax._src import abstract_arrays
-from jax._src import dtypes
 
 
 class Cauchy:
     def abstract_eval(self, key, shape=()):
         return (
             key,
-            abstract_arrays.ShapedArray(shape=shape, dtype=dtypes.float_),
+            abstract_arrays.ShapedArray(shape=shape, dtype=jnp.float32),
         )
 
     def sample(self, key, **kwargs):
